Se creo el paquete componentes
Dentro del paquete componentes se agrego el archivo Menu.js que facilita
la creación de menús desplegables para formularios.

Se extrajo el código del componente <Menu> de ReportesEstadisticas.js.
